Move server actions out of firebaseAdmin into their own file
